Move ServiceApp.*fields to controller
